+++ Assets/Scripts/Interfaces/IDamageable.cs
@@ -2,6 +2,6 @@ using UnityEngine;
 
 public interface IDamageable
 {
-    void Damage(float damage);
-    void Damage(float damage, Vector3 knockbackDirection);
+    void Damage(int damage);
+    void Damage(int damage, Vector3 knockbackDirection);
 }
